+++ scripts/store/configureStore.js
@@ -1,11 +1,5 @@
-import { createStore, applyMiddleware, compose } from "redux";
-import thunk from "redux-thunk";
-import rootReducer from "../reducers";
-
-const finalCreateStore = compose(
-  applyMiddleware(thunk)
-)(createStore);
-
-export default function configureStore(initialState) {
-  return finalCreateStore(rootReducer, initialState);
+if (process.env.NODE_ENV === "production") {
+  module.exports = require("./configureStore.production");
+} else {
+  module.exports = require("./configureStore.development");
 }
